UICONSET-162: Save, edit Authorization policies for consortium
